+++ src/de/ueller/midlet/gps/tile/PaintContext.java
@@ -104,7 +104,7 @@ public class PaintContext extends ScreenContext {
 	/** distance to next connection in meters when following the route path */
 	public volatile float conWayDistanceToNext;
 	/** time in 1/10 seconds to next connection when following the route path */
-	public volatile float conWayDurationToNext;
+	public volatile int conWayDurationToNext;
 	/** when painting draw highlighted path on top */
 	public volatile boolean highlightedPathOnTop;
 	/** used to find out if the connection leads to multiple same named ways*/
